energy_management.py
# energy_management.py

def collect_sensor_data():
    """
    Simulates collecting sensor data from IoT devices.
    Replace this with actual data collection logic.
    """
    sensor_data = [
        {"appliance": "Refrigerator", "energy_usage": 120},
        {"appliance": "Washing Machine", "energy_usage": 300},
        # Add more simulated sensor data or actual data collection logic
    ]
    return sensor_data

def analyze_energy_consumption(sensor_data):
    """
    Analyzes the collected sensor data for energy consumption patterns.
    """
    analyzed_data = {}  # Placeholder for analyzed data
    # Implement data analysis logic based on collected sensor_data
    # For instance, compute average usage, peak hours, etc.
    return analyzed_data

def schedule_energy_optimization(analyzed_data):
    """
    Schedules energy optimization based on analyzed data.
    """
    # Implement logic to schedule energy optimization actions
    # For example, schedule appliances during off-peak hours, etc.
    pass

# Main execution
if __name__ == "__main__":
    # Collect sensor data from IoT devices
    sensor_data = collect_sensor_data()

    # Analyze energy consumption patterns
    analyzed_data = analyze_energy_consumption(sensor_data)

    # Schedule energy optimization based on analysis
    schedule_energy_optimization(analyzed_data)
