test_energy_management.py
import unittest
from unittest.mock import patch
from scripts.energy_management import collect_sensor_data, analyze_energy_consumption, schedule_energy_optimization

class TestEnergyManagement(unittest.TestCase):

    @patch('scripts.energy_management.collect_sensor_data')
    def test_collect_sensor_data(self, mock_collect_sensor_data):
        # Mocking sensor data collection function
        mock_collect_sensor_data.return_value = [
            {"appliance": "Refrigerator", "energy_usage": 120},
            {"appliance": "Washing Machine", "energy_usage": 300}
            # Add more mocked sensor data if needed
        ]
        sensor_data = collect_sensor_data()
        self.assertEqual(len(sensor_data), 2)  # Ensure correct number of sensor data collected

    def test_analyze_energy_consumption(self):
        # Test analyze_energy_consumption function with dummy data
        sensor_data = [
            {"appliance": "Refrigerator", "energy_usage": 120},
            {"appliance": "Washing Machine", "energy_usage": 300}
            # Add more sensor data for testing analysis function
        ]
        analyzed_data = analyze_energy_consumption(sensor_data)
        # Add assertions for the analyzed data based on your logic

    def test_schedule_energy_optimization(self):
        # Test schedule_energy_optimization function (if applicable)
        # Add relevant test scenarios and assertions here
        pass

if __name__ == '__main__':
    unittest.main()
